+++ CloudShell.UI.Composition/CompositionAttributeNames.cs
@@ -1,0 +1,6 @@
+namespace CloudShell.UI.Composition;
+
+public static class CompositionAttributeNames
+{
+    public const string Icon = "cloudshell.ui.composition.icon";
+}